requirements.txt
aiofiles==25.1.0
    # via code-graph (pyproject.toml)
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via llama-index-core
aiosignal==1.4.0
    # via aiohttp
aiosqlite==0.21.0
    # via llama-index-core
annotated-doc==0.0.3
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
banks==2.2.0
    # via llama-index-core
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   nltk
    #   uvicorn
colorama==0.4.6
    # via griffe
cryptography==46.0.3
    # via
    #   pyjwt
    #   python-jose
dataclasses-json==0.6.7
    # via llama-index-core
deprecated==1.3.1
    # via
    #   banks
    #   llama-index-core
    #   llama-index-instrumentation
dirtyjson==1.0.8
    # via llama-index-core
ecdsa==0.19.1
    # via python-jose
fastapi==0.121.0
    # via code-graph (pyproject.toml)
filetype==1.2.0
    # via llama-index-core
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.10.0
    # via llama-index-core
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.28.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.187.0
    # via google-generativeai
google-auth==2.43.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.1
    # via google-api-python-client
google-generativeai==0.8.5
    # via
    #   code-graph (pyproject.toml)
    #   llama-index-embeddings-gemini
    #   llama-index-llms-gemini
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.2.4
    # via sqlalchemy
griffe==1.14.0
    # via banks
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   code-graph (pyproject.toml)
    #   llama-index-core
    #   mcp
    #   ollama
httpx-sse==0.4.3
    # via mcp
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via banks
joblib==1.5.2
    # via nltk
jsonschema==4.25.1
    # via mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
llama-index-core==0.14.7
    # via
    #   code-graph (pyproject.toml)
    #   llama-index-embeddings-gemini
    #   llama-index-embeddings-ollama
    #   llama-index-graph-stores-neo4j
    #   llama-index-llms-gemini
    #   llama-index-llms-ollama
llama-index-embeddings-gemini==0.4.1
    # via code-graph (pyproject.toml)
llama-index-embeddings-ollama==0.8.3
    # via code-graph (pyproject.toml)
llama-index-graph-stores-neo4j==0.5.1
    # via code-graph (pyproject.toml)
llama-index-instrumentation==0.4.2
    # via llama-index-workflows
llama-index-llms-gemini==0.6.1
    # via code-graph (pyproject.toml)
llama-index-llms-ollama==0.9.0
    # via code-graph (pyproject.toml)
llama-index-workflows==2.11.0
    # via llama-index-core
    # via code-graph (pyproject.toml)
loguru==0.7.3
    # via code-graph (pyproject.toml)
markupsafe==3.0.3
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
mcp==1.20.0
    # via code-graph (pyproject.toml)
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
neo4j==5.28.2
    # via
    #   code-graph (pyproject.toml)
    #   llama-index-graph-stores-neo4j
nest-asyncio==1.6.0
    # via llama-index-core
networkx==3.5
    # via llama-index-core
nltk==3.9.2
    # via llama-index-core
numpy==2.3.4
    # via llama-index-core
ollama==0.6.0
    # via
    #   llama-index-embeddings-ollama
    #   llama-index-llms-ollama
packaging==25.0
    # via
    #   marshmallow
    #   pytest
pillow==10.4.0
    # via
    #   llama-index-core
    #   llama-index-llms-gemini
platformdirs==4.5.0
    # via
    #   banks
    #   llama-index-core
pluggy==1.6.0
    # via pytest
prometheus-client==0.23.1
    # via code-graph (pyproject.toml)
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.12.4
    # via
    #   code-graph (pyproject.toml)
    #   banks
    #   fastapi
    #   google-generativeai
    #   llama-index-core
    #   llama-index-instrumentation
    #   llama-index-workflows
    #   mcp
    #   ollama
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.11.0
    # via
    #   code-graph (pyproject.toml)
    #   mcp
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via mcp
pyparsing==3.2.5
    # via httplib2
pytest==8.4.2
    # via pytest-asyncio
pytest-asyncio==1.2.0
    # via llama-index-embeddings-ollama
python-dotenv==1.2.1
    # via
    #   code-graph (pyproject.toml)
    #   pydantic-settings
    #   uvicorn
python-jose==3.5.0
    # via code-graph (pyproject.toml)
python-multipart==0.0.20
    # via
    #   code-graph (pyproject.toml)
    #   mcp
pytz==2025.2
    # via neo4j
pyyaml==6.0.3
    # via
    #   llama-index-core
    #   uvicorn
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.11.3
    # via
    #   nltk
    #   tiktoken
requests==2.32.5
    # via
    #   google-api-core
    #   llama-index-core
    #   tiktoken
rpds-py==0.28.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via
    #   google-auth
    #   python-jose
setuptools==80.9.0
    # via llama-index-core
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.44
    # via llama-index-core
sqlglot==27.29.0
    # via code-graph (pyproject.toml)
sse-starlette==3.0.3
    # via mcp
starlette==0.49.3
    # via
    #   fastapi
    #   mcp
tenacity==9.1.2
    # via llama-index-core
tiktoken==0.12.0
    # via llama-index-core
tqdm==4.67.1
    # via
    #   google-generativeai
    #   llama-index-core
    #   nltk
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   aiosqlite
    #   anyio
    #   fastapi
    #   google-generativeai
    #   grpcio
    #   llama-index-core
    #   llama-index-workflows
    #   pydantic
    #   pydantic-core
    #   pytest-asyncio
    #   referencing
    #   sqlalchemy
    #   starlette
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via
    #   dataclasses-json
    #   llama-index-core
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via
    #   code-graph (pyproject.toml)
    #   mcp
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
wrapt==2.0.0
    # via
    #   deprecated
    #   llama-index-core
yarl==1.22.0
    # via aiohttp
